Adding rma namespace and adding support for rma_object and is_rma_eligible traits
Moving lots of rma memory registration code into the rma namespace and
out of the parcelport subdirectories.

An rma_object is any type that has a contiguous memory layout and can in
principle be a source or destination of rma read/write type operations.

is_rma_eligible is a trait that is used to enable rma creation of a
type.

rma_object provides a wrapper (like shared_ptr) around a T, creation
of an rma_object must be via the make_rma_object(...) function which
will use the rma allocator provided by the parcelport layer (currently
the default parcelport, though this will be extended to take a locality
ID to ensure that the correct parcelport is used for a given
destination).

Currently rma_object supports only simple T types, vectors of T and
arrays of T.

Provide allocator support for STL (vector) using
rma_object<std::vector<T, rma_allocator<T>>>